Add Dutch translations for 0-setup.dns.md
Adds Dutch translations for 0-setup.dns.md.

This one was particularly hard to translate because a lot of the terminology concerning DNS records are in English and have no good translations to Dutch. (Not to my knowledge at least)
I apologise if some of the paragraphs are complete and utter nonsense to the people that happen to need this page of documentation.
Any remarks or corrections to this document are very welcome, I don't completely understand DNS records myself.